#12094: clean up moreh_adam
