+++ ttnn/cpp/ttnn/operations/moreh/moreh_adam/moreh_adam.cpp
@@ -21,7 +21,7 @@ std::vector<std::optional<Tensor>> MorehAdam::invoke(
     const std::optional<const Tensor> exp_avg_out,
     const std::optional<const Tensor> exp_avg_sq_out,
     const std::optional<const Tensor> max_exp_avg_sq_out,
-    const MemoryConfig& memory_config,
+    const std::optional<ttnn::MemoryConfig>& memory_config,
     const DeviceComputeKernelConfig compute_kernel_config) {
     return ttnn::prim::moreh_adam(
         param_in,
