+++ ttnn/cpp/ttnn/operations/moreh/moreh_pybind.hpp
@@ -3,11 +3,10 @@
 // SPDX-License-Identifier: Apache-2.0
 
 #pragma once
+#include "pybind11/pybind_fwd.hpp"
 
-#include <pybind11/pybind11.h>
-
-#include "ttnn/operations/moreh/moreh_adam/moreh_adam_pybind.hpp"
+namespace py = pybind11;
 
 namespace ttnn::operations::moreh {
-void py_module(py::module& module) { moreh_adam::bind_moreh_adam_operation(module); }
+void bind_moreh_operations(py::module& module);
 }  // namespace ttnn::operations::moreh
